rcn_django/requirements.txt
channels==3.0.4
channels-redis==3.3.1
Django==3.2.8
django-cors-headers==3.10.0
pubnub==4.8.1
python-decouple==3.5
requests==2.26.0
ringcentral==0.7.11

